+++ src/main/java/gdg/festa/application/service/sms/SmsCertifyService.java
@@ -1,11 +1,11 @@
-package gdg.festa.application.service;
+package gdg.festa.application.service.sms;
 
 import gdg.festa.application.usecase.sms.SmsCertifyUseCase;
 import gdg.festa.core.exception.CustomException;
 import gdg.festa.core.exception.ErrorCode;
 import gdg.festa.core.util.SmsUtil;
 import gdg.festa.infrastructure.redis.SmsCertification;
-import gdg.festa.presentation.request.SmsCertifyRequestDto;
+import gdg.festa.presentation.request.sms.SmsCertifyRequestDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
