+++ src/main/java/gdg/festa/application/usecase/sms/SmsCertifyUseCase.java
@@ -1,7 +1,7 @@
 package gdg.festa.application.usecase.sms;
 
 import gdg.festa.core.annotation.UseCase;
-import gdg.festa.presentation.request.SmsCertifyRequestDto;
+import gdg.festa.presentation.request.sms.SmsCertifyRequestDto;
 
 @UseCase
 public interface SmsCertifyUseCase {
