+++ src/main/java/gdg/festa/application/usecase/sms/SmsVertifyUseCase.java
@@ -1,7 +1,7 @@
 package gdg.festa.application.usecase.sms;
 
 import gdg.festa.core.annotation.UseCase;
-import gdg.festa.presentation.request.SmsVerifyRequestDto;
+import gdg.festa.presentation.request.sms.SmsVerifyRequestDto;
 
 @UseCase
 public interface SmsVertifyUseCase {
