+++ src/main/java/gdg/festa/presentation/controller/SmsController.java
@@ -3,8 +3,8 @@ package gdg.festa.presentation.controller;
 import gdg.festa.application.usecase.sms.SmsCertifyUseCase;
 import gdg.festa.application.usecase.sms.SmsVertifyUseCase;
 import gdg.festa.core.common.CommonResponseDto;
-import gdg.festa.presentation.request.SmsCertifyRequestDto;
-import gdg.festa.presentation.request.SmsVerifyRequestDto;
+import gdg.festa.presentation.request.sms.SmsCertifyRequestDto;
+import gdg.festa.presentation.request.sms.SmsVerifyRequestDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
