+++ src/main/java/gdg/festa/presentation/request/sms/SmsCertifyRequestDto.java
@@ -1,4 +1,4 @@
-package gdg.festa.presentation.request;
+package gdg.festa.presentation.request.sms;
 
 public record SmsCertifyRequestDto(
         String phoneNumber
