+++ src/main/java/gdg/festa/presentation/request/sms/SmsVerifyRequestDto.java
@@ -1,4 +1,4 @@
-package gdg.festa.presentation.request;
+package gdg.festa.presentation.request.sms;
 
 public record SmsVerifyRequestDto(
         String phoneNumber,
